[imporve] move ai package and improve ai code.
