+++ manager/src/main/java/org/apache/hertzbeat/manager/controller/AiController.java
@@ -22,8 +22,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.apache.hertzbeat.manager.config.AiProperties;
-import org.apache.hertzbeat.manager.service.AiService;
-import org.apache.hertzbeat.manager.service.impl.AiServiceFactoryImpl;
+import org.apache.hertzbeat.manager.service.ai.AiService;
+import org.apache.hertzbeat.manager.service.ai.factory.AiServiceFactoryImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.codec.ServerSentEvent;
 import org.springframework.util.Assert;
@@ -59,10 +59,8 @@ public class AiController {
     @Operation(summary = "Artificial intelligence questions and Answers",
             description = "Artificial intelligence questions and Answers")
     public Flux<ServerSentEvent<String>> requestAi(@Parameter(description = "Request text", example = "Who are you") @RequestParam("text") String text)  {
-
         Assert.notNull(aiServiceFactory, "please check that your type value is consistent with the documentation on the website");
         AiService aiServiceImplBean = aiServiceFactory.getAiServiceImplBean(aiProperties.getType());
-
         return aiServiceImplBean.requestAi(text);
     }
 }
