+++ manager/src/test/java/org/apache/hertzbeat/manager/service/AiServiceFactoryTest.java
@@ -24,7 +24,8 @@ import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import org.apache.hertzbeat.common.constants.AiTypeEnum;
-import org.apache.hertzbeat.manager.service.impl.AiServiceFactoryImpl;
+import org.apache.hertzbeat.manager.service.ai.AiService;
+import org.apache.hertzbeat.manager.service.ai.factory.AiServiceFactoryImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
